Refactor converter module
