+++ src/services/transactions_list.rs
@@ -2,7 +2,7 @@ extern crate reqwest;
 
 use crate::config::{base_transaction_service_url, request_cache_duration};
 use crate::models::backend::transactions::Transaction;
-use crate::models::service::transactions::list::TransactionSummary;
+use crate::models::service::transactions::summary::TransactionSummary;
 use crate::models::commons::Page;
 use crate::utils::context::Context;
 use crate::utils::extract_query_string;
@@ -23,7 +23,7 @@ pub fn get_all_transactions(context: &Context, safe_address: &String, next: &Opt
     println!("{:#?}", body);
     let backend_transactions: Page<Transaction> = serde_json::from_str(&body)?;
     let service_transactions: Vec<TransactionSummary> = backend_transactions.results.into_iter()
-        .flat_map(|transaction| transaction.to_service_transaction(&mut info_provider, safe_address).unwrap_or(vec!()))
+        .flat_map(|transaction| transaction.to_transaction_summary(&mut info_provider, safe_address).unwrap_or(vec!()))
         .collect();
 
     Ok(Page {
